fake_data_platform/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pandas
numpy
faker
python-dotenv
pydantic
APScheduler
fastapi
uvicorn[standard]
pyarrow       
APScheduler==3.10.4